digit_recognition/requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=9.0.0
flask>=2.0.0
opencv-python>=4.6.0